+++ pandas/tests/sparse/frame/test_frame.py
@@ -350,6 +350,20 @@ class TestSparseDataFrame(SharedWithSparse):
         assert sdf.default_fill_value == 0
         tm.assert_frame_equal(sdf.to_dense(), df)
 
+        @pytest.mark.filterwarnings("ignore:Sparse:FutureWarning")
+        def test_deprecated_to_sparse():
+            # GH 26557
+            # Deprecated 0.25.0
+
+            df = pd.DataFrame({"A": [1, np.nan, 3]})
+            sparse_df = pd.SparseDataFrame({"A": [1, np.nan, 3]})
+
+            # Deprecated 0.25.0
+            with tm.assert_produces_warning(FutureWarning,
+                                            check_stacklevel=False):
+                result = df.to_sparse()
+            tm.assert_frame_equal(result, sparse_df)
+
     def test_density(self):
         df = SparseSeries([nan, nan, nan, 0, 1, 2, 3, 4, 5, 6])
         assert df.density == 0.7
