+++ pandas/tests/sparse/series/test_series.py
@@ -1552,6 +1552,21 @@ def test_to_sparse():
 
 
 @pytest.mark.filterwarnings("ignore:Sparse:FutureWarning")
+def test_deprecated_to_sparse():
+    # GH 26557
+    # Deprecated 0.25.0
+    
+    ser = Series([1, np.nan, 3])
+    sparse_ser = pd.SparseSeries([1, np.nan, 3])
+    
+    with tm.assert_produces_warning(FutureWarning,
+                                    check_stacklevel=False):
+        result = ser.to_sparse()
+    tm.assert_series_equal(result, sparse_ser)
+
+
+
+@pytest.mark.filterwarnings("ignore:Sparse:FutureWarning")
 def test_constructor_mismatched_raises():
     msg = "Length of passed values is 2, index implies 3"
     with pytest.raises(ValueError, match=msg):
@@ -1576,4 +1591,4 @@ def test_deprecate_to_dense():
     # Deprecated 0.25.0
     with tm.assert_produces_warning(FutureWarning):
         result = sparse_ser.to_dense()
-        tm.assert_series_equal(result, ser)
+    tm.assert_series_equal(result, ser)
